pkl,pvm: a bunch of small fixes
2020-04-03  Jose E. Marchesi  <jemarch@gnu.org>

	* src/ios-dev-file.c (ios_dev_file_pread): Fix type of variable.
	(ios_dev_file_pwrite): Likewise.
	* src/pk-cmd.c (pk_cmd_exec_1): Likewise.
	* src/pvm.jitter (aref): Remove unneeded check.

+++ src/ios-dev-file.c
@@ -146,7 +146,7 @@ static int
 ios_dev_file_pread (void *iod, void *buf, size_t count, ios_dev_off offset)
 {
   struct ios_dev_file *fio = iod;
-  int ret;
+  size_t ret;
 
   /* We are using FILE* for buffering, rather than low-level fd, so we
      have to fake low-level pread by using fseeko.  */
@@ -166,7 +166,7 @@ ios_dev_file_pwrite (void *iod, const void *buf, size_t count,
                      ios_dev_off offset)
 {
   struct ios_dev_file *fio = iod;
-  int ret;
+  size_t ret;
 
   /* We are using FILE* for buffering, rather than low-level fd, so we
      have to fake low-level pread by using fseeko.  */
